+++ game-pcbuildingsimulator2/index.js
@@ -2,8 +2,8 @@
 Name: PC Building Simulator 2 Vortex Extension
 Structure: Unity BepinEx (IL2CPP)
 Author: ChemBoy1
-Version: 0.1.0
-Date: 2025-10-15
+Version: 0.1.1
+Date: 2026-03-16
 //////////////////////////////////////////*/
 
 //Import libraries
@@ -34,15 +34,21 @@ const GAME_REGSTRING = "PCBS2";
 const BEPINEX_PAGE_ID = '0'; //only specify if there is a Nexus page for BepInEx
 const BEPINEX_FILE_ID = '0';
 const BEPINEX_ARCH = 'x64'; // 'x64' or 'x86'
-const BEPINEX_BUILD = 'unityil2cpp'; // 'unityil2cpp' or 'unitymono' 
-const BEPINEX_VERSION = '6.0.0'; //force BepInEx version ('5.4.23.3' or '6.0.0')
+const BEPINEX_BUILD = 'il2cpp'; // 'il2cpp' or 'mono' 
+let BEPINEX_VERSION = '6.0.0'; //force BepInEx version ('5.4.23.5' or '6.0.0')
+const BEP_BE_VER = '755'; //set BepInEx build for BE URLs
+const BEP_BE_COMMIT = '3fab71a'; //git commit number for BE builds
+const BEPCFGMAN_VER = '18.4.1'; //set BepInExConfigManager version for direct URLs
 const allowBepinexNexus = false; //set false until bugs are fixed
 const downloadCfgMan = false; //should BepInExConfigManager be downloaded?
+if (BEPINEX_VERSION == '6.0.0') {
+  BEPINEX_VERSION = `${BEPINEX_VERSION}-be.${BEP_BE_VER}+${BEP_BE_COMMIT}`;
+}
 
-const BEPINEXIL2CPP_BE_URL = `https://builds.bepinex.dev/projects/bepinex_be/738/BepInEx-Unity.IL2CPP-win-x64-6.0.0-be.738%2Baf0cba7.zip`;
+const BEPINEXIL2CPP_BE_URL = `https://builds.bepinex.dev/projects/bepinex_be/${BEP_BE_VER}/BepInEx-Unity.IL2CPP-win-x64-6.0.0-be.${BEP_BE_VER}%2B${BEP_BE_COMMIT}.zip`;
 const BEPINEXIL2CPP_BE_URL_ERR = `https://builds.bepinex.dev/projects/bepinex_be`;
 const BEPINEX_ID = 'bepinex-injector';
-const BEPINEX_ZIP = 'BepInEx-Unity.IL2CPP-win-x64-6.0.0-be.738+af0cba7.zip';
+const BEPINEX_ZIP = `BepInEx-Unity.IL2CPP-win-x64-6.0.0-be.${BEP_BE_VER}+${BEP_BE_COMMIT}.zip`;
 
 let GAME_PATH = '';
 let STAGING_FOLDER = '';
@@ -54,14 +60,20 @@ const ROOT_ID = `${GAME_ID}-root`;
 const ROOT_NAME = "Root Game Folder";
 
 let BEPINEX_STRING = 'mono';
-if (BEPINEX_BUILD === 'unityil2cpp') {
+if (BEPINEX_BUILD === 'il2cpp') {
   BEPINEX_STRING = 'il2cpp';
 }
 const BEPCFGMAN_ID = `${GAME_ID}-bepcfgman`;
 const BEPCFGMAN_NAME = "BepInEx Configuration Manager";
 const BEPCFGMAN_PATH = 'Bepinex';
-const BEPCFGMAN_URL = `https://github.com/sinai-dev/BepInExConfigManager/releases/download/1.3.0/BepInExConfigManager.${BEPINEX_STRING}.zip`;
-const BEPCFGMAN_FILE = `bepinexconfigmanager.${BEPINEX_STRING}.dll`; //lowercased
+const BEPCFGMAN_FILE = `configurationmanager.dll`; //lowercased
+let BEPCFGMAN_ARCHIVE_NAME = `BepInEx.ConfigurationManager_BepInEx5_v`;
+let BEPCFGMAN_URL= `https://github.com/BepInEx/BepInEx.ConfigurationManager/releases/download/v${BEPCFGMAN_VER}/BepInEx.ConfigurationManager_BepInEx5_v${BEPCFGMAN_VER}.zip`;
+const BEPCFGMAN_URL_ERR = `https://github.com/BepInEx/BepInEx.ConfigurationManager/releases`;
+if (BEPINEX_BUILD === 'il2cpp') {
+  BEPCFGMAN_ARCHIVE_NAME = `BepInEx.ConfigurationManager_IL2CPP_v`;
+  BEPCFGMAN_URL = `https://github.com/BepInEx/BepInEx.ConfigurationManager/releases/download/v${BEPCFGMAN_VER}/BepInEx.ConfigurationManager_IL2CPP_v${BEPCFGMAN_VER}.zip`;
+}
 
 const BEPMOD_ID = `${GAME_ID}-bepmods`;
 const BEPMOD_NAME = "BepInEx Mod";
@@ -603,6 +615,13 @@ function applyGame(context, gameSpec) {
     const gameId = selectors.activeGameId(state);
     return gameId === GAME_ID;
   }); //*/
+  context.registerAction('mod-icons', 300, 'open-ext', {}, 'Download BepInExConfigManager', () => {
+    downloadBepCfgMan(context.api, spec);
+    }, () => {
+      const state = context.api.getState();
+      const gameId = selectors.activeGameId(state);
+      return gameId === GAME_ID;
+  });
   context.registerAction('mod-icons', 300, 'open-ext', {}, 'Open BepInEx.cfg', () => {
     GAME_PATH = getDiscoveryPath(context.api);
     const openPath = path.join(GAME_PATH, 'BepinEx', 'config', 'BepInEx.cfg');
@@ -656,9 +675,10 @@ function main(context) {
       context.api.ext.bepinexAddGame({
         gameId: GAME_ID,
         autoDownloadBepInEx: true,
+        architecture: BEPINEX_ARCH,
+        bepinexVersion: BEPINEX_VERSION,
         customPackDownloader: () => {
-          downloadBepinex(context.api, spec);
-          //return path.join(DOWNLOAD_FOLDER, BEPINEX_ZIP);
+          return downloadBepinex(context.api, spec);
         },
       });
     }
@@ -761,6 +781,7 @@ async function downloadBepCfgMan(api, gameSpec) {
       util.batchDispatch(api.store, batched); // Will dispatch both actions
     } catch (err) {
       api.showErrorNotification(`Failed to download/install ${MOD_NAME}`, err);
+      util.open(BEPCFGMAN_URL_ERR).catch(() => null);
     } finally {
       api.dismissNotification(NOTIF_ID);
     }