+++ game-discoelysium/CHANGELOG.md
@@ -4,6 +4,11 @@
 
 - None
 
+## [0.1.2] - 2026-03-16
+
+- Fixed: Bump BepInEx BE to version 755
+- Fixed: Migrated to more updated version of BepInExConfigManager
+
 ## [0.1.1] - 2026-01-27
 
 - Fixed a possible deployment error due to missing BepInEx patchers folder.