+++ game-pcbuildingsimulator2/CHANGELOG.md
@@ -4,6 +4,11 @@
 
 - None
 
+## [0.1.1] - 2026-03-16
+
+- Fixed: Bump BepInEx BE to version 755 and corrected displayed version
+- Added: Button to download BepInExConfigManager
+
 ## [0.1.0] - 2025-10-15
 
 - Initial release